+++ client/src/routes/Alerts/SELCustom.jsx
@@ -1,21 +1,11 @@
-import {
-  Button,
-  TableCell,
-  Tooltip,
-  Box,
-  LinearProgress,
-  Dialog,
-  DialogContent,
-} from "@mui/material"
-import { useState, useContext } from "react"
-import { apiConfig } from "../../config"
-import { UserContext } from "../../contexts/UserContext"
+import { Box, Button, Dialog, DialogContent, LinearProgress, TableCell, Tooltip } from "@mui/material"
+import { useContext, useState } from "react"
 
+import SELTable from "./SELTable"
+import { UserContext } from "../../contexts/UserContext"
+import { apiConfig } from "../../config"
 import { useQuery } from "react-query"
 import { useSnackbar } from "notistack"
-
-// FIXME: This is temporary
-import SELTable from "../Node/components/SELTable"
 
 const SELCustom = ({ data, node, type, icon }) => {
   const [openSEL, setOpenSEL] = useState(false)
@@ -33,9 +23,7 @@ const SELCustom = ({ data, node, type, icon }) => {
         },
         signal,
       }
-      const res = await (
-        await fetch(`${apiConfig.apiUrl}/redfish/v1/sel/${node}`, payload)
-      ).json()
+      const res = await (await fetch(`${apiConfig.apiUrl}/redfish/v1/sel/${node}`, payload)).json()
       if (res.status === "error") {
         enqueueSnackbar(res.message, { variant: "error" })
         setOpenSEL(false)
@@ -56,28 +44,22 @@ const SELCustom = ({ data, node, type, icon }) => {
   }
   let msg = ""
 
-  if (data[type] !== undefined && data[type].message !== null)
-    msg = ": " + data[type].message
+  if (data[type] !== undefined && data[type].message !== null) msg = ": " + data[type].message
   return (
     <>
       <TableCell>
         <Button variant="outlined" onClick={handleOpenSEL}>
           <Tooltip title={type + msg}>
-            <i
-              className={`bi ${icon}`}
-              style={{ color: iconColor(type), fontSize: "25px" }}
-            />
+            <i className={`bi ${icon}`} style={{ color: iconColor(type), fontSize: "25px" }} />
           </Tooltip>
         </Button>
       </TableCell>
       <Dialog open={openSEL} onClose={handleCloseSEL} maxWidth="xl">
         <DialogContent>
           {query_sel.isLoading && <LinearProgress />}
-          {!query_sel.isLoading &&
-            query_sel.status === "success" &&
-            query_sel.data.status === "success" && (
-              <SELTable data={query_sel.data.logs} />
-            )}
+          {!query_sel.isLoading && query_sel.status === "success" && query_sel.data.status === "success" && (
+            <SELTable data={query_sel.data.logs} />
+          )}
           <Box
             sx={{
               marginTop: "20px",
