+++ client/src/routes/Node/old_grendel/BootImage.jsx
