+++ client/src/routes/Node/old_redfish/Console.jsx
