+++ client/src/routes/Node/old_redfish/TerminalC.jsx
